Modificacion y terminacion de la libreria pwm . Cambio de metodos en libreria gpio en la carpeta de los pwm

+++ Code/Galileo/Perifericos/GPIO/gpio_robocol.h
@@ -208,7 +208,7 @@ gpio_st gpio_set_dir(uint8_t num,int esEntrada);
 **							requiera isntanciar el GPIO.
 */
 /* ===================================================================*/
-gpio_st gpio_gal_value(uint8_t num,int valor);
+gpio_st gpio_gal_value(uint8_t num,uint8_t valor);
 
 gpio_st gpio_gal_set(uint8_t num);
